Move protocols into `core`
